% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{mutate_east_north}
\alias{mutate_east_north}
\title{Title}
\usage{
mutate_east_north(wind_data)
}
\description{
Given a tibble with wind vector direction in degrees and the average signed
wind vector speed, return a tibble with the northerly and easterly wind components.
}
